Se corrige el nombre del proyecto
