+++ app/build.gradle
@@ -4,11 +4,11 @@ plugins {
 }
 
 android {
-    namespace 'com.example.vinilosgrupo4'
+    namespace 'com.example.vinilosgrupo15'
     compileSdk 32
 
     defaultConfig {
-        applicationId "com.example.vinilosgrupo4"
+        applicationId "com.example.vinilosgrupo15"
         minSdk 21
         targetSdk 32
         versionCode 1
