+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.VinilosGrupo4"
+        android:theme="@style/Theme.VinilosGrupo15"
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
