--- app/src/main/java/com/example/vinilosgrupo4/repository/BandRepository.kt
@@ -1,14 +1,0 @@
-package com.example.vinilosgrupo4.repository
-
-import com.example.vinilosgrupo4.servicesadapter.BandService
-import com.example.vinilosgrupo4.servicesadapter.RetrofitClient
-
-class BandRepository {
-    private var apiService: BandService? = null
-
-    init {
-        apiService = RetrofitClient.getClient?.create(BandService::class.java)
-    }
-
-    suspend fun getAllBands() = apiService?.getBands()
-}
