--- app/src/main/java/com/example/vinilosgrupo4/servicesadapter/BandService.kt
@@ -1,11 +1,0 @@
-package com.example.vinilosgrupo4.servicesadapter
-
-import com.example.vinilosgrupo4.model.BandReponseDataModel
-import retrofit2.Response
-import retrofit2.http.GET
-
-public interface BandService {
-    @GET("bands")
-
-    suspend fun getBands(): Response<ArrayList<BandReponseDataModel>>
-}
