--- app/src/main/java/com/example/vinilosgrupo4/view/MainActivity.kt
@@ -1,17 +1,0 @@
-package com.example.vinilosgrupo4.view
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import androidx.databinding.DataBindingUtil
-import com.example.vinilosgrupo4.R
-import com.example.vinilosgrupo4.databinding.ActivityMainBinding
-
-class MainActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityMainBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = DataBindingUtil.setContentView(this, R.layout.activity_main)
-        setContentView(binding.root)
-    }
-}
