--- app/src/main/java/com/example/vinilosgrupo4/view/viewholder/itemMusicianViewHolder.kt
@@ -1,22 +1,0 @@
-package com.example.vinilosgrupo4.view.viewholder
-
-import androidx.recyclerview.widget.RecyclerView
-import com.bumptech.glide.Glide
-import com.example.vinilosgrupo4.databinding.MusicianRowBinding
-import com.example.vinilosgrupo4.model.MusiciansResponseDataModel
-
-class itemMusicianViewHolder(binding: MusicianRowBinding): RecyclerView.ViewHolder(binding.root) {
-    private var binding: MusicianRowBinding? = null
-
-    init {
-        this.binding = binding
-    }
-
-    fun setItem(model: MusiciansResponseDataModel) {
-        binding?.let { view ->
-            view.name = model.name
-
-            Glide.with(view.root.context).load(model.image).into(view.imgCharacter)
-        }
-    }
-}
