+++ app/src/main/res/navigation/bandnavigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/bandListFragment"
-        android:name="com.example.vinilosgrupo4.view.fragments.BandListFragment"
+        android:name="com.example.vinilosgrupo15.view.fragments.BandListFragment"
         android:label="fragment_band_list"
         tools:layout="@layout/fragment_band_list" >
         <action
@@ -16,6 +16,6 @@
     </fragment>
     <fragment
         android:id="@+id/bandDetailFragment"
-        android:name="com.example.vinilosgrupo4.view.fragments.BandDetailFragment"
+        android:name="com.example.vinilosgrupo15.view.fragments.BandDetailFragment"
         android:label="BandDetailFragment" />
 </navigation>
